+++ application/Model/RunUnit/Email.php
@@ -79,7 +79,7 @@ class Email extends RunUnit {
         }
 
         if (!knitting_needed($this->body)) {
-            $this->body_parsed = $parsedown->text($this->body);
+            $this->body_parsed = parsedown_text_safe($parsedown, $this->body, 'the email body');
         }
 
         $this->db->insert_update('survey_emails', array(