+++ application/Model/RunUnit/Pause.php
@@ -53,7 +53,8 @@ class Pause extends RunUnit {
         $parsedown->setBreaksEnabled(true);
 
         if (!knitting_needed($this->body) && !empty($this->body)) {
-            $this->body_parsed = $parsedown->text($this->body); // transform upon insertion into db instead of at runtime
+            // transform upon insertion into db instead of at runtime
+            $this->body_parsed = parsedown_text_safe($parsedown, $this->body, 'the pause text');
         }
 
         $this->db->insert_update('survey_pauses', array(